Move Store and IAction to Store namespace
