+++ src/Store.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a02d75d158074e38827d00fb389218ad
+timeCreated: 1582479655
